+++ src/components/specific/layout/index.tsx
@@ -8,10 +8,10 @@
 import React from "react"
 import { useStaticQuery, graphql } from "gatsby"
 import Nav from "../Nav"
-import Header from "../../Header"
+import Header from "../Header"
 import ScrollToTop from "react-scroll-up"
 import styles from "./styles.module.scss"
-import IconBar from "../../IconBar"
+import IconBar from "../IconBar"
 import "bulma/css/bulma.css"
 
 interface Props {
